backend/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cryptography==45.0.4
decorator==5.2.1
defusedxml==0.7.1
Django==4.2.23
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf_base64==2.0
exceptiongroup==1.3.0
executing==2.2.0
fpdf==1.7.2
gunicorn==23.0.0
idna==3.10
ipython==8.18.1
isort==6.0.1
jedi==0.19.2
matplotlib-inline==0.1.7
oauthlib==3.2.2
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
reportlab==4.4.1
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
